refactor(shifu): isolate promotion command orchestration
